fix: fix logic when creating cart during registration

+++ src/main/java/com/musicstore/musicstoreapi/dto/response/productDTO/ProductResponse.java
@@ -4,6 +4,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @Builder
@@ -11,5 +13,5 @@ public class ProductResponse {
     private Integer id;
     private String name;
     private String imageUrl;
-    private Long price;
+    private BigDecimal price;
 }
